// src/app/modules/alarm/models/patientvisits.model.ts
import { Base } from "../../global/model/base.model";
export class PatientVisits extends Base {
  id: number = 0;
  ihe_msg_id: number = 0;
  patient_id: number = 0;
  visit_number: string = "";
  visited_on: Date = new Date();
  patient_info: string = "";
  patient_type_id: number = 0;
  patient_adm_class_id: number = 0;
  patient_adm_class: string = "";
  admission_dttm: Date = new Date();
  patient_financial_class: string = "";
  point_of_care: string = "";
  room: string = "";
  bed: string = "";
  floor: string = "";
  facility: string = "";
  building: string = "";
  patient_height: string = "";
  patient_height_uom: string = "";
  patient_weight: string = "";
  patient_weight_uom: string = "";
  diagnosis_code: string = "";
  attending_dr_code: string = "";
  attending_dr_family_name: string = "";
  attending_dr_given_name: string = "";
  attending_dr_assigning_authority_code: string = "";
  attending_dr_assigning_authority_type: string = "";
  admitting_dr_code: string = "";
  admitting_dr_family_name: string = "";
  admitting_dr_given_name: string = "";
  admitting_dr_assigning_authority_code: string = "";
  admitting_dr_assigning_authority_type: string = "";
  enterprise_id: number = 0;
  ent_location_id: number = 0;
  row_key: string = "";
  created_by: number = 0;
  modified_by: number = 0;
  created_on: Date = new Date();
  modified_on: Date = new Date();
  is_active: boolean = false;
  is_suspended: boolean = false;
  parent_id: number = 0;
  is_factory: boolean = false;
  notes: string = "";
  attributes: string = "";
  has_device_association: boolean = false;
}
export class PatientVisitsWrapper extends PatientVisits {
  patient_order_id: number = 0;
  order_code: string = "";
  order_status: string = "";
}